Mouse,Keyboard interaction reworked finished

+++ src/main/java/de/robinschleser/the12lords/gui/Position2D.java
@@ -3,28 +3,28 @@ package de.robinschleser.the12lords.gui;
 
 public class Position2D {
 
-    private int x, y;
+    private double x, y;
 
-    public Position2D(int x, int y) {
+    public Position2D(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
     public Position2D() {}
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
